Move code from __init__to a separate file
